price-optimiser-backend/requirements.txt
asgiref==3.8.1
dj-database-url==3.0.0
Django==5.2.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
numpy==2.2.6
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
whitenoise==6.9.0

# Machine Learning and Data Science dependencies
scikit-learn==1.5.2
pandas==2.2.3
matplotlib==3.9.1
seaborn==0.13.2
